+++ src/main/java/ru/innopolis/stc9/dao/interfaces/AttendanceDao.java
@@ -5,11 +5,11 @@ import ru.innopolis.stc9.pojo.Attendance;
 import java.util.List;
 
 public interface AttendanceDao {
-    boolean addLessonAttendance(int lessonId, int[] students);
+    boolean addLessonAttendance(long lessonId, long[] students);
 
-    int getNumberOfMissedLessons(int id);
+    int getNumberOfMissedLessons(long id);
 
     boolean updateAttendance(Attendance attendance);
 
-    List<Attendance> getLessonAttendance(int lessonId, int groupId);
+    List<Attendance> getLessonAttendance(long lessonId, long groupId);
 }
