+++ src/main/java/ru/innopolis/stc9/dao/interfaces/HomeWorkDao.java
@@ -12,17 +12,17 @@ public interface HomeWorkDao {
 
     boolean updateHomeWork(HomeWork homeWork);
 
-    HomeWork findById(int id);
+    HomeWork findById(long id);
 
-    HomeWork findByStudentId(int studentId);
+    HomeWork findByStudentId(long studentId);
 
-    HomeWork findByLessonId(int lessonId);
+    HomeWork findByLessonId(long lessonId);
 
-    List<HomeWork> getHomeWorkListByLessonId(int lessonId);
+    List<HomeWork> getHomeWorkListByLessonId(long lessonId);
 
-    HomeWork findHomeWorkByStudentIdAndLessonId(int studentId, int lessonId);
+    HomeWork findHomeWorkByStudentIdAndLessonId(long studentId, long lessonId);
 
     List<HomeWork> findAllHomeWork();
 
-    boolean deleteHomeWork(int id);
+    boolean deleteHomeWork(long id);
 }
