+++ src/main/java/ru/innopolis/stc9/service/implementation/ProgressServiceImpl.java
@@ -64,7 +64,7 @@ public class ProgressServiceImpl implements ProgressService {
 
     @Override
     public int getNumberOfMissedLessons(String login) {
-        int id = getUser(login).getId();
+        long id = getUser(login).getId();
         return attendanceService.getNumberOfMissedLessons(id);
     }
